+++ src/pages/403.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+const Forbidden = () => {
+  return (
+    <div>
+      The page you are looking for is not for your eyes
+    </div>
+  );
+}
+export default Forbidden;