+++ src/pages/analytics.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+const Analytics = () => {
+  return (
+    <div>
+      This is Analytics page. It will be implemented in V 2.0
+    </div>
+  );
+}
+export default Analytics;