+++ src/pages/listing.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+const Listing = () => {
+  return (
+    <div>
+      This is listing for all stock tickers
+    </div>
+  );
+}
+export default Listing;