+++ src/pages/singleStock.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+const SingleStock = () => {
+  return (
+    <div>
+      Single stock with all the information
+    </div>
+  );
+}
+export default SingleStock;