+++ src/pages/user.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+const UserPanel = () => {
+  return (
+    <div>
+      Change your user identity
+    </div>
+  );
+}
+export default UserPanel;